backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
numpy
torch
torchvision
transformers
scikit-learn
Pillow

